+++ apps/ai-dial-admin/src/components/Rules/Value/RulesValue.tsx
@@ -13,7 +13,7 @@ import { BASE_ICON_PROPS } from '@/src/constants/main-layout';
 import { useI18n } from '@/src/locales/client';
 import { DialRule, RuleFunction } from '@/src/models/dial/rule';
 import { DropdownItemsModel } from '@/src/models/dropdown-item';
-import { getAttributeItems, getOperationItems } from './rules';
+import { getAttributeItems, getOperationItems } from '../utils';
 
 interface Props {
   rule: DialRule;
